% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-trail.R
\name{audit_log_set_event_type}
\alias{audit_log_set_event_type}
\title{Set Audit Log Event Type}
\usage{
audit_log_set_event_type(event_type, req)
}
\arguments{
\item{event_type}{The event type to be set for the audit log.}

\item{req}{The request object, which should contain an audit log in its internal data.}
}
\value{
Returns nothing as it modifies the audit log in-place.
}
\description{
This function sets the event type for an audit log. It retrieves the audit log from the request's
internal data, and then calls the audit log's set_event_type method with the provided event type.
}
