% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-trail.R
\name{audit_log_set_study_id}
\alias{audit_log_set_study_id}
\title{Set Audit Log Study ID}
\usage{
audit_log_set_study_id(study_id, req)
}
\arguments{
\item{study_id}{The study ID to be set for the audit log.}

\item{req}{The request object, which should contain an audit log in its internal data.}
}
\value{
Returns nothing as it modifies the audit log in-place.
}
\description{
This function sets the study ID for an audit log. It retrieves the audit log from the request's
internal data, and then calls the audit log's set_study_id method with the provided study ID.
}
