% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-trail.R
\name{AuditLog}
\alias{AuditLog}
\title{AuditLog Class}
\description{
This class is used internally to store audit logs for each request.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuditLog-new}{\code{AuditLog$new()}}
\item \href{#method-AuditLog-disable}{\code{AuditLog$disable()}}
\item \href{#method-AuditLog-is_enabled}{\code{AuditLog$is_enabled()}}
\item \href{#method-AuditLog-set_request_body}{\code{AuditLog$set_request_body()}}
\item \href{#method-AuditLog-set_response_body}{\code{AuditLog$set_response_body()}}
\item \href{#method-AuditLog-set_ip_address}{\code{AuditLog$set_ip_address()}}
\item \href{#method-AuditLog-set_user_agent}{\code{AuditLog$set_user_agent()}}
\item \href{#method-AuditLog-set_event_type}{\code{AuditLog$set_event_type()}}
\item \href{#method-AuditLog-set_study_id}{\code{AuditLog$set_study_id()}}
\item \href{#method-AuditLog-set_response_code}{\code{AuditLog$set_response_code()}}
\item \href{#method-AuditLog-validate_log}{\code{AuditLog$validate_log()}}
\item \href{#method-AuditLog-persist}{\code{AuditLog$persist()}}
\item \href{#method-AuditLog-clone}{\code{AuditLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$new(request_method, endpoint_url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-disable"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-disable}{}}}
\subsection{Method \code{disable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$disable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-is_enabled"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-is_enabled}{}}}
\subsection{Method \code{is_enabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$is_enabled()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_request_body"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_request_body}{}}}
\subsection{Method \code{set_request_body()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_request_body(request_body)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_response_body"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_response_body}{}}}
\subsection{Method \code{set_response_body()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_response_body(response_body)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_ip_address"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_ip_address}{}}}
\subsection{Method \code{set_ip_address()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_ip_address(ip_address)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_user_agent"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_user_agent}{}}}
\subsection{Method \code{set_user_agent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_user_agent(user_agent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_event_type"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_event_type}{}}}
\subsection{Method \code{set_event_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_event_type(event_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_study_id"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_study_id}{}}}
\subsection{Method \code{set_study_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_study_id(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-set_response_code"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-set_response_code}{}}}
\subsection{Method \code{set_response_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$set_response_code(response_code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-validate_log"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-validate_log}{}}}
\subsection{Method \code{validate_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$validate_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-persist"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-persist}{}}}
\subsection{Method \code{persist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$persist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuditLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuditLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuditLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
