% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-minimisation-pocock.R
\name{compare_rows}
\alias{compare_rows}
\title{Compare rows of two dataframes}
\usage{
compare_rows(all_patients, new_patients)
}
\arguments{
\item{all_patients}{data.frame with all patients}

\item{new_patients}{data.frame with new patient}
}
\value{
data.frame with columns as in all_patients and new_patients,
filled with TRUE if there is match in covariate and FALSE if not
}
\description{
Takes dataframe all_patients (presumably with one row / patient) and
compares it to all rows of new_patients (presumably already randomized
patients)
}
