% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_db_connection_pool}
\alias{create_db_connection_pool}
\title{Defines methods for interacting with the study in the database
Create a database connection pool}
\usage{
create_db_connection_pool(...)
}
\value{
A pool object representing the connection pool to the database.
}
\description{
This function creates a connection pool to a PostgreSQL database. It uses
environment variables to get the necessary connection parameters. If the
connection fails, it will retry up to 5 times with a delay of 2 seconds
between each attempt.
}
\examples{
\dontrun{
pool <- create_db_connection_pool()
}
}
