% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-define.R
\name{define_study}
\alias{define_study}
\title{Define study}
\usage{
define_study(
  name,
  identifier,
  arms,
  method = c("simple", "block"),
  strata = list(),
  parameters = NULL,
  ratio = rep(1, times = length(arms))
)
}
\arguments{
\item{name}{\code{character(1)}\cr
Full study name.}

\item{identifier}{\code{character(1)}\cr
Study code, at most 12 characters.}

\item{arms}{\code{character()}\cr
Arm names to use.}

\item{method}{\code{character(1)}\cr
Randomization method to apply.}

\item{strata}{\code{list()}\cr
List of character vectors, each list element being a stratum and each string
being a possible stratum value. Could possibly take a numeric structure as
well instead of a character vector, e.g. \code{list(min = 1, max = 10)}. It just
needs handling by checking whether the inner list is named or not, I'd say.}

\item{parameters}{\code{list()}\cr
Parameters to pass to randomization.}

\item{ratio}{\code{integer()}\cr
Arm ratios, must be positive and the same length as arm names.}
}
\value{
This function is called for the side effect of updating the DB.
}
\description{
Creates a study with specified parameters and publishes it to the DB.
}
\examples{
\dontrun{
define_study(
  "DEMO", "Demonstrational study", c("placebo", "active"),
  method = "simple",
  strata = list(gender = c("F", "M"), working = c("yes", "no"))
)
}

}
