% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-repository.R
\name{get_similar_studies}
\alias{get_similar_studies}
\title{Defines methods for interacting with the study in the database}
\usage{
get_similar_studies(name, identifier)
}
\description{
Defines methods for interacting with the study in the database
}
