% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-list.R
\name{list_studies}
\alias{list_studies}
\title{List available studies}
\usage{
list_studies()
}
\value{
A tibble with basic study info, including ID.
}
\description{
Queries the DB for the basic information about existing studies.
}
