% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-dynamic.R
\name{randomize_dynamic}
\alias{randomize_dynamic}
\title{Randomize Dynamic Algorithm for Patient Allocation}
\usage{
randomize_dynamic(
  arms,
  current_state,
  weights,
  ratio,
  method = "var",
  p = 0.85
)
}
\arguments{
\item{arms}{\code{character()}\cr
Arm names.}

\item{current_state}{\code{data.frame()}\cr
table of covariates and current arm assignments in column \code{arm},
last row contains the new patient with empty string for \code{arm}}

\item{weights}{\code{numeric()}\cr
vector of positive weights, equal in length to number of covariates,
numbered after covariates, defaults to equal weights}

\item{ratio}{\code{integer()}\cr
Vector of positive integers, equal in length to number of arms,
named after arms, defaults to equal weight}

\item{method}{\code{character()}\cr
Function used to compute within-arm variability, must be one of:
\code{sd}, \code{var}, \code{range}, defaults to \code{var}}

\item{p}{\code{numeric()}\cr
single value, proportion of randomness (0, 1) in the randomization
vs determinism, defaults to 85\% deterministic}
}
\value{
\code{character()}\cr
name of the arm assigned to the patient
}
\description{
The \code{randomize_dynamic} function implements the dynamic randomization
algorithm using the minimization method proposed by Pocock (Pocock and Simon
1975). It requires defining basic study parameters: the number of arms (k),
covariate values, patient allocation ratios (\(a_{k}\)), weights for the
covariates (\(w_{i}\)), and the maximum probability of assigning a patient to
the group with the smallest total unbalance multiplied by the respective
weights (\(G_{k}\)). As the total unbalance for the first patient is the same
regardless of the assigned arm, this patient is randomly allocated to a given
arm. Subsequent patients are randomized based on the calculation of the
unbalance depending on the selected method: "range", "var" (variance), or
"sd" (standard deviation). In the case of two arms, the "range" method is
equivalent to the "sd" method.
}
\details{
Initially, the algorithm creates a matrix of results comparing a newly
randomized patient with the current balance of patients based on the defined
covariates (C). In the next step, for each arm and specified covariate,
various scenarios of patient allocation are calculated. The existing results
(n) are updated with the new patient, and then, considering the ratio
coefficients, the results are divided by the specific allocation ratio
(\(a_{k}\)). Depending on the method, the total unbalance is then calculated,
taking into account the allocation (\(a_{k}\)) and the number of covariates,
where i = 1,2,…,C.
\itemize{
\item \code{range}: \(G_{k} = \ \sum_{i = 1}^{c}w_{i}\lbrack RANGE(\frac{n_{ir_{i}1}}{a_{1}},\frac{n_{ir_{i}1}}{a_{2}},\ldots,\ \frac{n_{ir_{i}k}}{a_{k}})|\),
\item \code{var}: \(G_{k} = \ \sum_{i = 1}^{c}w_{i}\lbrack VAR(\frac{n_{ir_{i}1}}{a_{1}},\frac{n_{ir_{i}1}}{a_{2}},\ldots,\ \frac{n_{ir_{i}k}}{a_{k}})|\),
\item \code{sd}: \(G_{k} = \ \sum_{i = 1}^{c}w_{i}\lbrack SD(\frac{n_{ir_{i}1}}{a_{1}},\frac{n_{ir_{i}1}}{a_{2}},\ldots,\ \frac{n_{ir_{i}k}}{a_{k}})|\)
}

Based on the number of defined arms (K), the minimum value of \(G_{k}\)
(defined as the weighted sum of the level-based imbalance) selects the arm to
which the patient will be assigned with a predefined probability. The
probability that a patient will be assigned to any other arm will then be
\(\frac{(1 - p)}{(K - 1)}\) for each of the remaining arms.
}
\examples{
n_at_the_moment <- 10
arms <- c("control", "active low", "active high")
sex <- sample(c("F", "M"),
              n_at_the_moment + 1,
              replace = TRUE,
              prob = c(0.4, 0.6)
)
diabetes <-
  sample(c("diabetes", "no diabetes"),
         n_at_the_moment + 1,
         replace = TRUE,
         prob = c(0.2, 0.8)
  )
arm <-
  sample(arms,
         n_at_the_moment,
         replace = TRUE,
         prob = c(0.4, 0.4, 0.2)
  ) |>
  c("")
covar_df <- data.frame(sex, diabetes, arm)
covar_df

randomize_dynamic(arms = arms, current_state = covar_df)

}
\references{
Pocock, S. J., & Simon, R. (1975). Minimization: A new method of assigning patients to treatment and control groups in clinical trials.
}
