% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-simple.R
\name{randomize_simple}
\alias{randomize_simple}
\title{Simple randomization}
\usage{
randomize_simple(arms, ratio)
}
\arguments{
\item{arms}{\code{character()}\cr
Arm names.}

\item{ratio}{\code{numeric()}\cr
Ratio of patient assignment to each arm. Must be the same length as \code{arms}.}
}
\value{
Selected arm assignment.
}
\description{
Randomly assigns a patient to one of the arms according to specified ratios,
regardless of already performed assignments.
}
\examples{
randomize_simple(c("active", "placebo"), c(2, 1))

}
