% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-details.R
\name{read_study_details}
\alias{read_study_details}
\title{Read study details}
\usage{
read_study_details(study_id)
}
\arguments{
\item{study_id}{\code{integer(1)}\cr
ID of the study.}
}
\value{
A tibble with study details, containing potentially complex columns,
like \code{arms}.
}
\description{
Queries the DB for the study parameters, including declared arms and strata.
}
