% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_api.R
\name{run_unbiased}
\alias{run_unbiased}
\title{Run API}
\usage{
run_unbiased(host = "0.0.0.0", port = 3838, ...)
}
\arguments{
\item{host}{\code{character(1)}\cr
Host URL.}

\item{port}{\code{integer(1)}\cr
Port to serve API under.}
}
\value{
Function called to serve the API in the caller thread.
}
\description{
Starts \pkg{unbiased} API.
}
