% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_db.R
\name{run_unbiased_db}
\alias{run_unbiased_db}
\title{Run local DB}
\usage{
run_unbiased_db()
}
\value{
Return code describing success or lack thereof.
}
\description{
Starts a Docker container containing a Postgres database for unbiased API
storage and sets environment variables to allow API to connect. Do not run
if an external Postgres database exists already, set appropriate env vars
instead.
}
