% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-trail.R
\name{setup_audit_trail}
\alias{setup_audit_trail}
\title{Set up audit trail}
\usage{
setup_audit_trail(pr)
}
\arguments{
\item{pr}{A plumber router for which the audit trail is to be set up.}
}
\value{
Returns the updated plumber router with the audit trail hooks.
}
\description{
This function sets up an audit trail for a given process. It uses plumber's hooks to log
information before routing (preroute) and after serializing the response (postserialize).
}
\details{
This function modifies the plumber router in place and returns the updated router.

The audit trail is only enabled if the AUDIT_LOG_ENABLED environment variable is set to "true".
}
\examples{
pr <- plumber::plumb("your-api-definition.R") |>
   setup_audit_trail()
}
