% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{setup_sentry}
\alias{setup_sentry}
\title{setup_sentry function}
\usage{
setup_sentry()
}
\arguments{
\item{None}{}
}
\value{
None. If the SENTRY_DSN environment variable is not set, the function will
return a message and stop execution.
}
\description{
This function is used to configure Sentry, a service for real-time error tracking.
It uses the sentryR package to set up Sentry based on environment variables.
}
\details{
The function first checks if the SENTRY_DSN environment variable is set. If not, it
returns a message and stops execution.
If SENTRY_DSN is set, it uses the sentryR::configure_sentry function to set up Sentry with
the following parameters:
\itemize{
\item dsn: The Data Source Name (DSN) is retrieved from the SENTRY_DSN environment variable.
\item app_name: The application name is set to "unbiased".
\item app_version: The application version is retrieved from the GITHUB_SHA environment variable.
If not set, it defaults to "unspecified".
\item environment: The environment is retrieved from the SENTRY_ENVIRONMENT environment variable.
If not set, it defaults to "development".
\item release: The release is retrieved from the SENTRY_RELEASE environment variable.
If not set, it defaults to "unspecified".
}
}
\examples{
setup_sentry()

}
\seealso{
\url{https://docs.sentry.io/}
}
