% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-list.R
\name{study_exists}
\alias{study_exists}
\title{Validate study existence}
\usage{
study_exists(study_id)
}
\arguments{
\item{study_id}{\code{integer(1)}\cr
ID of the study.}
}
\value{
\code{TRUE} or \code{FALSE}, depending whether given ID exists in the DB.
}
\description{
Checks the database for the existence of given ID.
}
