% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-minimisation-pocock.R
\name{compare_rows}
\alias{compare_rows}
\title{Compare rows of two dataframes}
\usage{
compare_rows(A, B)
}
\arguments{
\item{A}{data.frame with all patients}

\item{B}{data.frame with new patient}
}
\value{
data.frame with columns as in A and B, filled with TRUE if there is
match in covariate and FALSE if not
}
\description{
Takes dataframe B (presumably with one row / patient) and compares it to all
rows of A (presumably already randomized patietns)
}
