% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize-dynamic.R
\name{randomize_dynamic}
\alias{randomize_dynamic}
\title{Dynamic randomization}
\usage{
randomize_dynamic(
  arms,
  current_state,
  weights,
  ratio,
  method = "var",
  p = 0.85
)
}
\arguments{
\item{arms}{\code{character()}\cr
Arm names.}

\item{current_state}{\code{data.frame()}\cr
table of covariates and current arm assignments in column \code{arm}}

\item{weights}{\code{numeric()}\cr
vector of positive weights, equal in length to number of covariates,
numbered after covariates, defaults to equal weights}

\item{ratio}{\code{integer()}\cr
Vector of positive integers, equal in length to number of arms,
named after arms, defaults to equal weight}

\item{method}{\code{character()}\cr
Function used to compute within-arm variability, must be one of:
\code{sd}, \code{var}, \code{range}, defaults to \code{var}}

\item{p}{\code{numeric()}\cr
single value, proportion of randomness (0, 1) in the randomization
vs determinism, defaults to 85\% deterministic}
}
\value{
\code{character()}\cr
name of the arm assigned to the patient
}
\description{
Dynamic randomization
}
\examples{
n_at_the_moment <- 3
arms <- c("control", "active low", "active high")
sex <- sample(seq(1, 0),
  3,
  replace = TRUE,
  prob = c(0.4, 0.6)
)
diabetes <- sample(c("diabetes", "no diabetes"), nsample, replace = TRUE, prob = c(0.2, 0.8))
mat_of_covars <- cbind(c1, c2)
colnames(mat_of_covars) <- c("Sex", "Diabetes")
covar_class <- c("ordinal", "numeric", "ordinal", "numeric")
wght <- c(1 / 4, 1 / 4, 1 / 4, 1 / 4)

resrand <- integer()
init_pat <- length(c1)
resrand[1:init_pat] <- sample(arms,
  init_pat,
  replace = TRUE,
  prob = ratio / sum(ratio)
)

randomize_dynamic(
  covariates = mat_of_covars,
  patnum = init_pat + 1,
  weights = wght,
  ratio = ratio,
  no_of_trt = no_of_arms,
  arms = arms,
  current_state = resrand,
  p = 0.85,
  init_patnum = init_pat
)

}
