% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-trail.R
\name{setup_audit_trail}
\alias{setup_audit_trail}
\title{Set up audit trail}
\usage{
setup_audit_trail(pr, endpoints = list())
}
\arguments{
\item{pr}{A plumber router for which the audit trail is to be set up.}

\item{endpoints}{A list of regex patterns for which the audit trail should be enabled.}
}
\value{
Returns the updated plumber router with the audit trail hooks.
}
\description{
This function sets up an audit trail for a given process. It uses plumber's hooks to log
information before routing (preroute) and after serializing the response (postserialize).
}
\details{
This function modifies the plumber router in place and returns the updated router.
}
\examples{
pr <- plumber::plumb("your-api-definition.R") |>
  setup_audit_trail()
}
