% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased-package.R
\docType{package}
\name{unbiased-package}
\alias{unbiased}
\alias{unbiased-package}
\title{unbiased: Unbiased: Production-Grade Randomization API}
\description{
The Unbiased package delivers a minimization-based randomization algorithm for patient allocation in clinical trials, fully integrated with a production-ready API. It's designed to work seamlessly with a persistent PostgreSQL database, ensuring reliable data management and integrity. Packaged into precompiled Docker images, Unbiased simplifies deployment to just running docker-compose up, making it exceptionally straightforward to incorporate into your environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ttscience.github.io/unbiased/}
}

}
\author{
\strong{Maintainer}: Kamil Sijko \email{kamil.sijko@ttsi.com.pl} (\href{https://orcid.org/0000-0002-2203-1065}{ORCID})

Authors:
\itemize{
  \item Kinga Sałata \email{kinga.salata@ttsi.com.pl}
  \item Aleksandra Duda \email{aleksandra.duda@ttsi.com.pl}
  \item Łukasz Wałejko \email{lukasz.walejko@ttsi.com.pl}
  \item Jagoda Głowacka-Walas \email{jagoda.glowacka-walas@ttsi.com.pl} (\href{https://orcid.org/0000-0002-7628-8691}{ORCID})
  \item Laura Bąkała
}

Other contributors:
\itemize{
  \item Michał Seweryn \email{michal.seweryn@biol.uni.lodz.pl} (\href{https://orcid.org/0000-0002-9090-3435}{ORCID}) [contributor]
  \item Transition Technologies Science Sp. z o.o. [funder, copyright holder]
}

}
\keyword{internal}
