% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased-package.R
\docType{package}
\name{unbiased-package}
\alias{unbiased}
\alias{unbiased-package}
\title{unbiased: Diverse Randomization Algorithms for Clinical Trials}
\description{
The Unbiased package offers a comprehensive suite of randomization algorithms for clinical trials, encompassing dynamic strategies like the minimization method, simple randomization approaches, and block randomization techniques. Its primary purpose is to provide a harmonized set of functions that will seamlessly integrate with a production-ready plumber API, also contained within the package. This integration is designed to facilitate a smooth and efficient interface with electronic Case Report Form (eCRF) systems, enhancing the capability of clinical trials to manage patient allocation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ttscience.github.io/unbiased/}
}

}
\author{
\strong{Maintainer}: Kamil Sijko \email{kamil.sijko@ttsi.com.pl} (\href{https://orcid.org/0000-0002-2203-1065}{ORCID})

Authors:
\itemize{
  \item Kinga Sałata \email{kinga.salata@ttsi.com.pl}
  \item Aleksandra Duda \email{aleksandra.duda@ttsi.com.pl}
  \item Łukasz Wałejko \email{lukasz.walejko@ttsi.com.pl}
  \item Jagoda Głowacka-Walas \email{jagoda.glowacka-walas@ttsi.com.pl} (\href{https://orcid.org/0000-0002-7628-8691}{ORCID})
  \item Laura Bąkała
}

Other contributors:
\itemize{
  \item Michał Seweryn \email{michal.seweryn@biol.uni.lodz.pl} (\href{https://orcid.org/0000-0002-9090-3435}{ORCID}) [contributor]
  \item Transition Technologies Science Sp. z o.o. [funder, copyright holder]
}

}
\keyword{internal}
